refactoring hex2bin

+++ sha3dll/sha3.cpp
@@ -24,13 +24,13 @@ std::vector<sha3::bit> sha3::convertStringToBits(const std::vector<unsigned char
 	for (unsigned char c : str)
 	{
 		bits.push_back(bit((c & 0x80) == 0 ? 0 : 1));
-		bits.push_back(bit(((int)c & 0x40) == 0 ? 0 : 1));
-		bits.push_back(bit(((int)c & 0x20) == 0 ? 0 : 1));
-		bits.push_back(bit(((int)c & 0x10) == 0 ? 0 : 1));
-		bits.push_back(bit(((int)c & 0x08) == 0 ? 0 : 1));
-		bits.push_back(bit(((int)c & 0x04) == 0 ? 0 : 1));
-		bits.push_back(bit(((int)c & 0x02) == 0 ? 0 : 1));
-		bits.push_back(bit(((int)c & 0x01) == 0 ? 0 : 1));
+		bits.push_back(bit((c & 0x40) == 0 ? 0 : 1));
+		bits.push_back(bit((c & 0x20) == 0 ? 0 : 1));
+		bits.push_back(bit((c & 0x10) == 0 ? 0 : 1));
+		bits.push_back(bit((c & 0x08) == 0 ? 0 : 1));
+		bits.push_back(bit((c & 0x04) == 0 ? 0 : 1));
+		bits.push_back(bit((c & 0x02) == 0 ? 0 : 1));
+		bits.push_back(bit((c & 0x01) == 0 ? 0 : 1));
 	}
 
 	return std::move(bits);
